requirements.txt
bunch==1.0.1
colorlog==3.1.4
fasttext==0.8.3
gensim==3.5.0
hanja==0.11.0
jamo==0.4.1
JPype1==0.6.3
konlpy==0.4.4
matplotlib==2.2.2
numpy==1.15.0
pandas==0.23.3
pyahocorasick==1.1.8
scikit-learn==0.19.2
seaborn==0.9.0
sentencepiece==0.1.3
smart-open==1.6.0
soynlp==0.0.46
tensorboard==1.9.0
tensorflow-gpu==1.9.0
termcolor==1.1.0
tqdm==4.24.0
xgboost==0.72.1